refactor: GearChart から ChartHeader / PackTabSection を抽出
JSX 肥大化を解消するため、自己完結した 2 ブロックを独立コンポーネントに切り出し:

- ChartHeader.tsx (新規): 折りたたみトグル + Pie/Bar 表示モード切替
- PackTabSection.tsx (新規): ALL タブ + Pack タブ + 歯車ボタン + New Pack 入力
  (ref / state / focus effect を内部に内包)

GearChart.tsx:
- 1032 → 866 行 (約 170 行削減)
- showNewPackInput / newPackName / newPackInputRef / handleCreateNewPack を削除
  (PackTabSection 内部に移管)
- 親で必要な副作用 (Pack 切替時の選択/カテゴリクリア) は onPackChange コールバックに抽出

残課題 (次回 PR): Pie チャート本体 (weight-class 二重ドーナツ + 通常 Pie) を ChartBody に切り出す
